requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
transformers
pydantic
python-dotenv